# Moving attachment related attributes to attribs file
