+++ CSharp/Library/Microsoft.Bot.Builder/FormFlow/Attributes.cs
@@ -33,8 +33,10 @@
 
 using System;
 using System.Linq;
+using System.Threading.Tasks;
 
 using Microsoft.Bot.Builder.FormFlow.Advanced;
+using Microsoft.Bot.Connector;
 
 namespace Microsoft.Bot.Builder.FormFlow
 {
@@ -877,6 +879,60 @@ namespace Microsoft.Bot.Builder.FormFlow.Advanced
             LastSeparator = null;
             Separator = null;
             ValueCase = CaseNormalization.Default;
+        }
+    }
+
+    /// <summary>
+    /// Abstract base class used for attachment validation.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Field, AllowMultiple = true)]
+    public abstract class AttachmentValidatorAttribute : Attribute
+    {
+        public string ErrorMessage { get; set; }
+
+        public FormConfiguration Configuration { get; internal set; }
+
+        public abstract Task<bool> IsValidAsync(Attachment attachment, out string errorMessage);
+
+        public abstract string ProvideHelp();
+    }
+
+    /// <summary>
+    /// Attachment content-type validator attribute.
+    /// </summary>
+    public class AttachmentContentTypeValidatorAttribute : AttachmentValidatorAttribute
+    {
+        public AttachmentContentTypeValidatorAttribute() { }
+
+        public string ContentType { get; set; }
+
+        public override Task<bool> IsValidAsync(Attachment attachment, out string errorMessage)
+        {
+            errorMessage = default(string);
+            var result = attachment.ContentType.ToLowerInvariant().Contains(this.ContentType.ToLowerInvariant());
+
+            if (!result)
+            {
+                var template = this.Configuration.Template(TemplateUsage.AttachmentContentTypeValidatorError);
+
+                errorMessage = !string.IsNullOrWhiteSpace(this.ErrorMessage)
+                    ? this.ErrorMessage
+                    : string.Format(
+                        template.Pattern(),
+                        attachment.Name,
+                        string.IsNullOrWhiteSpace(this.ContentType) ? string.Empty : this.ContentType.ToLowerInvariant());
+            }
+
+            return Task.FromResult(result);
+        }
+
+        public override string ProvideHelp()
+        {
+            var template = this.Configuration.Template(TemplateUsage.AttachmentContentTypeValidatorHelp);
+
+            var contentType = string.IsNullOrWhiteSpace(this.ContentType) ? string.Empty : this.ContentType.ToLowerInvariant();
+
+            return string.Format(template.Pattern(), contentType);
         }
     }
 }
