+++ CSharp/Library/Microsoft.Bot.Builder/Microsoft.Bot.Builder.csproj
@@ -135,8 +135,6 @@
     <Compile Include="Dialogs\IDialog.cs" />
     <Compile Include="Dialogs\IDialogContext.cs" />
     <Compile Include="Base\Tasks.cs" />
-    <Compile Include="FormFlow\AttachmentContentTypeValidatorAttribute.cs" />
-    <Compile Include="FormFlow\AttachmentValidatorAttribute.cs" />
     <Compile Include="FormFlow\AwaitableAttachment.cs" />
     <Compile Include="FormFlow\FieldReflector.cs" />
     <Compile Include="Base\Traits.cs" />
